configs/beaglev_ahead: new defconfig
This patch adds board support for the BeagleV-Ahead. It uses the 6.
mainline kernel for the board's SoC.

The boot command for the custom uboot to boot the kernel could not
easily be changed. It seemed that a default boot command was persisent
and any change to it by normal means (e.g., uboot config parameters
like CONFIG_BOOT_COMMAND) did not have any effect. Hence a patch for
the custom u-boot is introduced for now, to insert the proper boot
command.

https://www.beagleboard.org/boards/beaglev-ahead